Move pod status functionality to pod::status
