+++ crates/kubelet/src/state.rs
@@ -3,10 +3,10 @@ use log::{debug, error, warn};
 
 pub mod prelude;
 
-use crate::pod::make_registered_status;
+use crate::pod::{initialize_pod_container_statuses, patch_status};
 use crate::pod::{Phase, Pod};
 use k8s_openapi::api::core::v1::Pod as KubePod;
-use kube::api::{Api, PatchParams};
+use kube::api::Api;
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
@@ -235,32 +235,6 @@ pub trait State<PodState>: Sync + Send + 'static + std::fmt::Debug {
     ) -> anyhow::Result<serde_json::Value>;
 }
 
-async fn patch_status(api: &Api<KubePod>, name: &str, patch: serde_json::Value) {
-    match serde_json::to_vec(&patch) {
-        Ok(data) => {
-            debug!(
-                "Applying status patch to Pod {}: '{}'",
-                &name,
-                std::str::from_utf8(&data).unwrap()
-            );
-            match api.patch_status(&name, &PatchParams::default(), data).await {
-                Ok(_) => (),
-                Err(e) => {
-                    warn!("Pod {} error patching status: {:?}", name, e);
-                }
-            }
-        }
-        Err(e) => {
-            warn!(
-                "Pod {} error serializing status patch {:?}: {:?}",
-                name, &patch, e
-            );
-        }
-    }
-}
-
-const MAX_STATUS_INIT_RETRIES: usize = 5;
-
 /// Iteratively evaluate state machine until it returns Complete.
 pub async fn run_to_completion<PodState: Send + Sync + 'static>(
     client: &kube::Client,
@@ -276,70 +250,11 @@ pub async fn run_to_completion<PodState: Send + Sync + 'static>(
         (name, api)
     };
 
-    // NOTE: This loop patches the container statuses of the Pod with and then
-    // waits for them to be picked up by the reflector. This is needed for a
-    // few reasons:
-    // * Kubernetes rewrites an empty array to null, preventing us from
-    //   starting with that and appending.
-    // * Pod reflection is not updated within a given state, meaning that
-    //   container status patching cannot be responsible for initializing this
-    //   (this would be a race condition anyway).
-    // I'm not sure if we want to loop forever or handle some sort of failure
-    // condition (if Kubernetes refuses to accept and propagate this
-    // initialization patch.)
-    let mut retries = 0;
-    'main: loop {
-        if retries == MAX_STATUS_INIT_RETRIES {
-            let patch = serde_json::json!(
-                {
-                    "metadata": {
-                        "resourceVersion": "",
-                    },
-                    "status": {
-                        "phase": Phase::Failed,
-                        "reason": "Timed out while initializing container statuses.",
-                    }
-                }
-            );
-            patch_status(&api, &name, patch).await;
-            return;
-        }
-        let (num_containers, num_init_containers) = {
-            let pod = pod.read().await;
-            patch_status(&api, &name, make_registered_status(&pod)).await;
-            let num_containers = pod.containers().len();
-            let num_init_containers = pod.init_containers().len();
-            (num_containers, num_init_containers)
-        };
-        for _ in 0..10 {
-            let status = {
-                pod.read()
-                    .await
-                    .as_kube_pod()
-                    .status
-                    .clone()
-                    .unwrap_or_default()
-            };
-
-            let num_statuses = status
-                .container_statuses
-                .as_ref()
-                .map(|statuses| statuses.len())
-                .unwrap_or(0);
-            let num_init_statuses = status
-                .init_container_statuses
-                .as_ref()
-                .map(|statuses| statuses.len())
-                .unwrap_or(0);
-
-            if (num_statuses == num_containers) && (num_init_statuses == num_init_containers) {
-                break 'main;
-            } else {
-                debug!("Pod {} waiting for status to populate: {:?}", &name, status);
-                tokio::time::delay_for(std::time::Duration::from_secs(1)).await;
-            }
-        }
-        retries += 1;
+    if initialize_pod_container_statuses(&name, Arc::clone(&pod), &api)
+        .await
+        .is_err()
+    {
+        return;
     }
 
     let mut state: Box<dyn State<PodState>> = Box::new(state);
